requirements.txt
typer==0.12.*
fastapi==0.115.*
uvicorn==0.30.*
GitPython==3.1.*
Jinja2==3.1.*
